bottomバー切り出し
